pyproject.toml
[project]
name = "lumina"
version = "1.0.0"
description = "LUMINA™ — The AI Visibility Platform. Be the answer."
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.34.0",
    "pydantic>=2.10.0",
    "httpx>=0.28.0",
    "anthropic>=0.42.0",
    "openai>=1.60.0",
    "google-genai>=1.5.0",
    "neo4j>=5.27.0",
    "pinecone-client>=5.1.0",
    "mcp>=1.3.0",
    "beautifulsoup4>=4.12.0",
    "tiktoken>=0.8.0",
    "numpy>=2.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=6.0.0",
    "ruff>=0.9.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/lumina"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"

[tool.ruff]
target-version = "py312"
line-length = 100
